backend/requirements.txt
django>=3.2,<4.0
djangorestframework>=3.14
django-cors-headers>=4.0
google-generativeai
numpy
